# openapi.yaml
openapi: 3.1.0
info:
  title: Archivault API
  version: 0.1.0
  description: |
    Cold storage SaaS API. Files are stored in AWS S3 Glacier Deep Archive (cold)
    or S3 Standard-IA (hot). All authenticated endpoints require a Clerk session token
    in the `Authorization: Bearer <token>` header.

    ## Upload flow
    1. `POST /api/files` → receive `uploadTarget.uploadUrl` (pre-signed S3 PUT URL)
    2. Client PUTs the file directly to S3 using that URL
    3. `PATCH /api/files/{fileId}` with `{ "action": "confirm_upload" }` to mark file active

    ## Retrieval flow
    1. `POST /api/retrievals` → kicks off Glacier restore job (12–48 hr)
    2. Poll `GET /api/retrievals` until status is `ready`
    3. Use `downloadUrl` from the retrieval record (valid for 1 hour)
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-domain.com
    description: Production
security:
  - clerkBearer: []
...
paths:
...
  /api/import/google:
    get:
      operationId: listGoogleDriveFiles
      tags: [Import]
      summary: List files and folders in the user's Google Drive
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: string
          description: Google Drive folder ID (omit for root)
      responses:
        '200':
          description: Array of Drive items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DriveItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Google Drive not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: enqueueGoogleDriveImport
      tags: [Import]
      summary: Enqueue selected Drive files for import into a vault
      description: |
        Creates a `files` record with status `importing` for each selected file and
        enqueues a QStash background job that streams the file from Google Drive to S3
        Glacier Deep Archive.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleImportRequest'
      responses:
        '201':
          description: Jobs queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/import/onedrive:
    get:
      operationId: listOneDriveItems
      tags: [Import]
      summary: List files and folders in the user's OneDrive
      parameters:
        - name: itemId
          in: query
          required: false
          schema: { type: string }
          description: OneDrive item ID (omit for root)
      responses:
        '200':
          description: Array of OneDrive items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OneDriveItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: OneDrive not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: enqueueOneDriveImport
      tags: [Import]
      summary: Enqueue selected OneDrive items for import into a vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneDriveImportRequest'
      responses:
        '201':
          description: Jobs queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneDriveImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/billing/checkout:
    post:
      operationId: createCheckoutSession
      tags: [Billing]
      summary: Create a Stripe checkout session to subscribe to a paid plan
      description: |
        Returns a Stripe-hosted checkout URL. Redirect the user to this URL.
        On success Stripe fires the `checkout.session.completed` webhook which
        updates the user's plan in the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutRequest'
      responses:
        '200':
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectUrlResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/webhooks/import-google:
    post:
      operationId: processGoogleDriveImportJob
      tags: [Webhooks]
      summary: Internal — stream a single Google Drive file to S3 Glacier (QStash signed)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileId, googleDriveFileId, userId]
              properties:
                fileId:
                  type: string
                googleDriveFileId:
                  type: string
                userId:
                  type: string
      responses:
        '200':
          description: Import job succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/webhooks/import-onedrive:
    post:
      operationId: processOneDriveImportJob
      tags: [Webhooks]
      summary: Internal — stream a single OneDrive file to S3 Glacier (QStash signed)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileId, onedriveItemId, userId]
              properties:
                fileId:
                  type: string
                onedriveItemId:
                  type: string
                userId:
                  type: string
      responses:
        '200':
          description: Import job succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    clerkBearer:
      type: http
      scheme: bearer
      description: Clerk session token (JWT)
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DriveItem:
      type: object
      required: [id, name, mimeType, sizeBytes, isFolder, modifiedAt]
      properties:
        id:
          type: string
          description: Google Drive file/folder ID
        name:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
          description: 0 for folders
        isFolder:
          type: boolean
        modifiedAt:
          type: string
          format: date-time
    OneDriveItem:
      type: object
      required: [id, name, mimeType, sizeBytes, isFolder, modifiedAt]
      properties:
        id:
          type: string
          description: OneDrive item ID
        name:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
          description: 0 for folders
        isFolder:
          type: boolean
        modifiedAt:
          type: string
          format: date-time
    OneDriveImportRequest:
      type: object
      required: [vaultId, items]
      properties:
        vaultId:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OneDriveItem'
    OneDriveImportResponse:
      type: object
      required: [queued]
      properties:
        queued:
          type: array
          items:
            type: object
            required: [fileId, onedriveItemId]
            properties:
              fileId:
                type: string
              onedriveItemId:
                type: string
    GoogleImportRequest:
      type: object
      required: [vaultId, files]
      properties:
        vaultId:
          type: string
          description: Target vault ID
        files:
          type: array
          description: Drive file items to import (folders are ignored)
          items:
            $ref: '#/components/schemas/DriveItem'
    GoogleImportResponse:
      type: object
      required: [queued]
      properties:
        queued:
          type: array
          items:
            type: object
            required: [fileId, driveFileId]
            properties:
              fileId:
                type: string
              driveFileId:
                type: string
...
    Plan:
      type: string
      enum: [free, starter, personal, creator, power]
    CreateCheckoutRequest:
      type: object
      required: [plan]
      properties:
        plan:
          $ref: '#/components/schemas/Plan'
          description: The paid plan to subscribe to (free is not valid here)
    RedirectUrlResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: Stripe-hosted redirect URL